Moved the location where get_entanglemet_entropy is defined
